requirements.txt
flask
python-dotenv
google-genai
pillow
gTTS